+++ src/strategy/hill_climb.rs
@@ -257,6 +257,9 @@ impl<G: HillClimbGenotype, F: Fitness<Genotype = G>, SR: StrategyReporter<Genoty
             None
         }
     }
+    fn flush_reporter(&mut self, output: &mut Vec<u8>) {
+        self.reporter.flush(output);
+    }
 }
 impl<G: HillClimbGenotype, F: Fitness<Genotype = G>, SR: StrategyReporter<Genotype = G>>
     HillClimb<G, F, SR>