+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.1] - 2024-09-17
+
+### Added
+* Add `StrategyReporterBuffer` implementation with an internal buffer instead of stdout
+* Add `flush_reporter()` in `Strategy` as the strategy can be boxed and we need
+  a way to get to the buffer of the `Reporter`
+
 ## [0.15.0] - 2024-09-17
 
 ### Changed